Make timescale test setup part of generic testing code
