+++ docker/docker-compose.test.yml
@@ -17,6 +17,7 @@ services:
 
   timescale:
     image: timescale/timescaledb:latest-pg12
+    command: postgres -F
     environment:
       POSTGRES_PASSWORD: 'postgres'
